Implemented a basic read loop
unfortunately there a bunch of bugs due to how fragile our parser is

e.g. typing ) will fuck it up with a segfault
e.g. typing (askdj will cause a panic (because we haven't handled errors gracefully in our parser)

What's worse is that when we encounter a parse error, we need to take care to clean up the memory allocated for the partially created SNode